Extract keys to its own trait
